Document VideoMotion class
- Rename video_sensing/lib.js to video_sensing/library.js
